kill one specific zombie by id (#19)
There is a list containing all the entities in the world. However, if we kill or pick up one entity, this entity will be removed from this list. In this case, isDead() cannot be used to judge on time.
However, if we spawn one entity by updating class entity into the world, we can set/get unique ID for it. In addition, class world provides us a method called getEntityById (disadvantage: this method only accepts integer, but the id created by getUniqueId is UUID. Maybe we can override this method but it is very complicated since its superclass is World.)
The solution is my code.

* add background & battle a monster
* add background & battle one monster
* add background & battle one monster
* kill zombie by id