Update the CraftItemGoal, program can check the item in player's inventory (#22)
Update the CraftItemGoal, TutorialMission and InventoryHandler with proper codes and comments. Add the prompts to instruct the player to make a wooden axe and now that the program about TutorialMission is ready to check whether the player make a wooden axe successfully by inspecting whether itemStack is in player's inventory.