Adding first version of run_session (#50)
The script run_session is ready to go, but it does not actually work that well,
and likely cannot work very well until changes are in the C++ code which we need
anyway. Some issues are listed in bugs.txt. Still, there is no harm playing with
run_session, and if things hang, there is the handy script kill_minecraftng,
also being introduced with this merge. 

This branch also includes some work in making scripts more robust, such as not
relying on TOMCAT variable. It is actually OK to rely on such a thing, but then
the script *must* test it is set to something reasonable and give an error
message and exit with failure if not the case. However, I prefer that the
scripts that users might run are able to sort out TOMCAT. It you put it in your
bashrc file, then it is wrong if you branch, etc.