FFmpeg integration, data collection streamlining, script refactoring (#66)
ffmpeg integration, data collection streamlining, script refactoring
================================================

This PR implements ffmpeg integration to record video of the player's face using the built-in webcam, video of the Minecraft window, and audio of the player while they are playing the game. It also streamlines data collection a bit and implements some refactoring of the scripts.

FFmpeg integration
--------------------

### Background

The three aforementioned input streams are essential for the ToMCAT research effort. Video of the player's face will be analyzed for affect, video of the Minecraft window will be used to establish human observer baselines and maybe in the service of retrospective self-report, and the audio recording will be used as input to the natural language dialogue pipeline.

In this implementation, we use ffmpeg to collect all three streams.

#### Original Malmo implementation of screen recording

Project Malmo has also implemented the functionality to record video of the Minecraft window. Basically, Minecraft sends individual RGB frames captured from the game to Malmo (see `VideoProducerImplementation.java`), that then timestamps them (with the timestamps for the frames stored in a separate text file), and then concatenates them into a video using `ffmpeg` with the `image2pipe` format (see `PosixFrameWriter.cpp`)

#### Why we are doing it differently

Turning on this feature significantly decreases the framerate at which the game runs. This is not a problem for the original envisioned purpose of Malmo - in which the 'avatars' were primarily controlled by AI agents, in the service of reinforcement learning. However, in our research, we are primarily concerned with gathering data from humans, and the reduced framerate leads to perceptible choppiness and a generally unpleasant experience for a human player.

### Caveats

In order to get a reasonable framerate for the output video of the screen recording, we focus ffmpeg on the Minecraft window, and crop out the rest of the screen. To do this, we obtain the position and size of the window using AppleScript (`tools/get_minecraft_window_position_and_size.scpt`), and process them to provide ffmpeg with the information necessary to crop the recording correctly. Of course, AppleScript is restricted to macOS machines only, so screen recording with this method has not been implemented for Linux yet. However, I believe similar functionality should be able to be achieved using the `wmctrl` tool on Linux - this is on the todo list for later.

Another caveat is that once ffmpeg has received the position and size of the Minecraft window, it will record the rectangle on the screen corresponding to that position and size. This means that if the Minecraft window is moved after this recording has started, the screen recording will not capture the whole window.

Data collection streamlining
----------------------------

The data collection process has been streamlined somewhat - now, the four data files produced in each session (player audio, player video, screen recording, Malmo data) are all placed in a timestamped directory inside the `data/participant_data` directory. The name of the directory is of the form `session_<year>_<month>_<day>_<hour>_<minute>_<second>`, corresponding to the time of creation of the directory.

Script refactoring
------------------

- Made the code that deals with setting the `TOMCAT` environment variable more concise.
- The code related to running the tutorial has been refactored out into a separate script, `run_tutorial`.

Miscellaneous
--------------

- Many .cpp, .h, and .java files have been autoformatted using `clang-format`.
- Some header include guards have been replaced with `#pragma once`.
- Some class method calls have been prefixed with `this->` to show explicitly that they are methods of the current class.
- Removed some namespace qualifiers.