Full screen ffmpeg recording (#67)
Since it turns out the experiments will be conducting with Minecraft running full screen, we don't really need to crop the ffmpeg recording area after all. This PR removes the code to find out the cropping bounds and replaces it with code to bring the Minecraft window to the foreground and make it full screen.

It also adds a flag to ffmpeg to make sure that the screen recording is not choppy when full screen, but stays at 30 fps.