Discrete Events (Button Press Event) (#68)
This PR adds functionality to detect discrete events. It specifically adds functionality to detect when a player presses a button, but it also adds the `ModBlocks` class, which can register any other new `Blocks` in the future by adding the relevant lines for the block. The class file contains more detailed information on how to do this which can be used to add new detector blocks easily.

The button event is only triggered when the player presses the new button added by this PR called the `ASIST Button`. It can be found in the Minecraft Creative menu under the Redstone tab, but it currently lacks a texture, so it shows up as a black and purple cube both in the inventory and in the world (after it is placed). 

When the button is pressed, a Hashmap output is printed to the terminal containing information about what the event was, the coordinates of the event, the time of occurrence (system date and time), and the player that caused the event.

NOTE: The current world on Vanga uses regular Minecraft buttons, not the new buttons added in this PR, so pressing buttons in the current version of the world will not print any output. 

NOTE: The button has defined borders (the same as normal Minecraft buttons) within which the player must right-click to register a press. Since the Block lacks a texture at the moment, it can be slightly confusing to understand where to click because the edges of the black and purple cube texture fall outside the defined boundaries; hence they don't register as a press.

The PR also implements auto formatting of all Java source code files with `clang-format`