Making ffmpeg testing more robust (#89)
The run_session.sh script now checks the error logs for specific error messages, and exits if the test audio and video files cannot be created, and the error messages are not found in the log file. The complete error log is printed to screen if the log file does not contain one of the known errors.