Adding list of potentially useful Minecraft events (#104)
This PR adds a lightly-curated list of discrete events that are available on the Minecraft Forge (v 1.11.2) event bus that could be useful for ToMCAT research activities. We can look over these and decide which ones we want to prioritize implementing subscriptions to. (Thanks to @Adi-UA for compiling the initial list).