Additional events: Lever flip, door open, block break, left click (#126)
Additional events: Lever flip, door open, block break, left clicks
=============================================

This PR implements a few additional events: lever flips, door opening, block breaking, and left clicking - to support the USAR single player mission.

A new field, `block_material` has been added to the data structure, and the `event_type` field now reflects the fully-qualified name of the corresponding event class (to accurately capture hypernymy relations between event types).

Examples
----------

### Block interaction

Event representing a block interaction (will be dispatched to different message bus topics depending on whether it was a left click or right click).

```json
{
    "player_name": "tomcat",
    "block_position": {
        "x": -2183.0,
        "y": 53.0,
        "z": 177.0
    },
    "block_type": "net.minecraft.block.BlockButtonStone",
    "block_material": "minecraft:stone_button",
    "event_type": "edu.arizona.tomcat.Events.BlockInteraction",
    "timestamp": "2020-04-26T23:11:22:515Z"
}
```

### Door interaction

The initial state of the door (i.e. before the player right-clicks on it) is captured with the `was_open` boolean field.

```json
{
    "was_open": false,
    "player_name": "tomcat",
    "block_position": {
        "x": -2153.0,
        "y": 53.0,
        "z": 176.0
    },
    "block_type": "net.minecraft.block.BlockDoor",
    "block_material": "minecraft:wooden_door",
    "event_type": "edu.arizona.tomcat.Events.DoorInteraction",
    "timestamp": "2020-04-26T23:11:28:560Z"
}
```

### Lever flip

The `was_powered` boolean field gives the state of the lever just before the player right-clicks it. It is a faithful representation of the underlying Minecraft lever block property. However, the actual semantics of this field are a bit context-sensitive. In the Zombie invasion mission, when a lever is powered, it points down (negative Y direction), but in the USAR singleplayer mission, when a lever is powered, it points up (positive Z direction). Even more confusingly, when the lever becomes powered in the Zombie mission, the iron door opens, however in the USAR Singleplayer mission, the lights in the hallway turn on when the lever is _unpowered_. So this is just something to keep in mind, and perhaps we want to separate out the event handlers a bit - rather than having most of them in `ForgeEventHandlers`, we may want to move some of the mission-specific event handling logic into the respective mission classes.

```json
{
    "was_powered": true,
    "player_name": "tomcat",
    "block_position": {
        "x": -2151.0,
        "y": 53.0,
        "z": 177.0
    },
    "block_type": "net.minecraft.block.BlockLever",
    "block_material": "minecraft:lever",
    "event_type": "edu.arizona.tomcat.Events.LeverFlip",
    "timestamp": "2020-04-26T23:11:31:213Z"
}
```

### Block break

The block break event was added to capture when victims are triaged in the USAR Singleplayer mission. Here is an example.

```json
{
    "player_name": "tomcat",
    "block_position": {
        "x": -2148.0,
        "y": 52.0,
        "z": 180.0
    },
    "block_type": "net.minecraft.block.Block",
    "block_material": "minecraft:gold_block",
    "event_type": "edu.arizona.tomcat.Events.BlockBreakEvent",
    "timestamp": "2020-04-26T23:12:00:510Z"
}
```

The type of victim is given by the block type and material (gold for severely injured victims, prismarine for less-severely injured victims).

Data that we may want to explicitly capture in the future include:
- type of click (left or right)
- item that was used to break the block
- start and stop times for durations when the player is breaking a block.