villager_saved event and refactoring to hit_controlled_door_opened (#134)
This PR adds the `villager_saved` event with information about the villager's `UUID` and `Position` (in addition to the `timestamp` and `eventType`). 

It also refactors `iron_door_opened` to `hit_controlled_door_opened`.

The `villager_saved` event is checked in a new method called `checkExtraEvents()` which has been added to the `ForgeEventHandler` class and is called every tick from `TomcatDecoratorImplementation`.

Also, in the `ZombieMission` class  the variable `MAX_DISTANCE_TO_SAVE_VILLAGER` has been made public, and a getter has been added for the `numberOfVillagersSaved`.