Revert "Updated textures (#137)" (#141)
This reverts commit adf155427731f7d7998b36181063471a4dd9a40b.

This reverts commit adf155427731f7d7998b36181063471a4dd9a40b.

My reasoning is that the higher-resolution textures are computationally more expensive and can impact the performance of the game, especially if we have three ffmpeg processes running at the same time. - so I don't want to have them enabled by default for all users. With the helper script introduced in #139, it should be straightforward for people to download and install the textures by doing the following:

- Downloading the textures from https://www.curseforge.com/minecraft/texture-packs/soartex-fanver-vanilla/download/2371741 and place the `Soartex_Fanver` folder in the `tomcat` directory.
- Running the command `./tools/install_soartex_fanver`