Multi client threads (#143)
This PR fixes the issues with double progress bar and double discrete event triggering in the zombie mission. This is accomplished by unregistering the classes we set as listeners of Minecraft events upon the end of the tutorial mission (it's implemented in a more generic way as in the end of any mission). The new mission will register again and only events related to it will be captured.