Possible fix for timestamp sorting (#168)
This PR implements a fix for a few corner cases where the elkless_replayer script did not sort the messages correctly by time. The issue was that the timestamps were being treated as strings, and the string comparison function did not work perfectly for comparing timestamps in this representation.

This update uses the `python-dateutil` package to parse the ISO-8601 timestamps properly and enable correct comparison. Thanks to @pmdoll for discovering this issue.