Client-side encryption (#171)
This PR (thanks to @chinmaib) implements client-side encryption of session data before uploading to the Vision server. The encryption can be enabled by setting the environment variable `ENCRYPT_DATA` to `1`.

Additionally, a new directory for configuration files, `conf` is introduced, and the scripts look for `netrc` and `rsa_public.pem` in there instead of the root directory.

The macOS portions of `tools/install_dependencies` has been streamlined a bit, and `openssl` has been added as a dependency.