YAML parsing to get spec version for metadata (#180)
This PR adds an executable to parse `docs/spec.yml` to extract the version number (currently 0.1) and add it to the `metadata.json` file produced by `tools/generate_session_metadata`. ToMCAT developers are advised to run `tools/install` after pulling this update, in order to install the `yaml-cpp` library and build the `getSpecVersion` executable.