Removing support for Ubuntu 18.04
This commit removes support for automated installation of dependencies
and continuous integration for Ubuntu 18.04 - we require libmosquitto
1.5 or later, while the version available using apt on Ubuntu 18.04 is
1.4.15. Additionally, this should speed up continuous integration since
we would no longer have to install a number of libraries from source,
since the required versions would be available through apt.

We also add an option to skip automatic dependency installation, for
users on operating systems other than macOS and Ubuntu 20.04 who are
able to install the requirements themselves. Setting the environment
variable INSTALL_DEPENDENCIES to 0 will skip the automated installation
of dependencies (it is set to 1 by default).