audioStreamer: Message bus and .wav saving implementation (#315)
Updates the audioStreamer with the functionality discussed in: #289

**MQTT Integration:**
The audioStreamer can now be triggered to start and stop streaming using an MQTT message bus. Streaming will start when the audioStreamer receives a "trial start" message and will stop with a "trial stop" message.

The audioStreamer does not need to be restarted between trials. When a "trial stop" message is received, it will reset and wait for another "trial start" message.

**Audio Saving:**
The audioStreamer can now save recorded audio locally to a .wav file. When not using MQTT mode, the filename will look like:
`$PLAYERNAME.wav`.

When using MQTT mode, to prevent file overwriting, the filename will  be generated using metadata  in the "trial start" message. The format will look like: 
`Trial-$TRIAL_Team-$TEAM_Member-$MEMBER.wav`.


**New command line options**:
1. use_mqtt(bool): Flag that determines if the audioStreamer will use MQTT for starting/stopping streaming
2. mqtt_host(string): The host that the MQTT broker is running on
3. mqtt_port(string): The port that the MQTT broker is running on
4. save_audio(bool): Flag that determines if the audioStreamer will save audio chunks to a .wav file
5. recordings_directory(string): The directory that audio recordings will be saved

**Additional changes**
- Separates components into separate source files.
- Adds paho.mqtt.cpp and libsndfile libraries as dependencies
- Improves error handling. 
- Adds logging.
 
Co-authored-by: Vincent Raymond <vraymond6@arizona.edu>