ASIST Study-3 preregistration (#331)
This PR adds the ASIST Study-3 preregistration document for the ToMCAT project. It reflects what was uploaded to the ASIST registry.

The document may undergo further edits just so that we have a nicer document for internal purposes, but for now I think it's ready to be transferred to the master branch.