In fNIRS_plot.py, replaced inlet.pull_sample() with inlet.pull_chuck().
Now the visualization program gets all the samples accumulated in LSL
while it was plotting the samples from the previous chunk.

The visualization program only plots the most recent sample in a chunk
discarding all the samples it could not get to in time while it was
plotting the last sample of the previous chunk.

This seem to have fixed the lagged plotting behavior in the
visualization program.