Modifications to start_lab_recorder script and config files (#448)
Modified start_lab_recorder:
Added a var to identify the "TARGET_IP" and set it to "cat.local". This var will be used as the IP address in all the ssh commands in the script. Also, I found by setting it to "cat.local" allows the script to run OK on TOM and CAT.
Added a section to run "git checkout <LabRecorder config files" after the 3 instances of LabRecorder are open and they are using the config files that have been modified with the current experiment data directory. This will put the config files back to the default way they are stored in the master git repository so that they will not look like they have been modified to other development branches (less merge issues to resolve).

Modified the default "StorageLocation=" setting in all 3 LabRecorder config files to be a more realistic, workable and an identifiable path. This is incase LabRecorder is opened without the start_lab_recorder script executing correctly and not getting the config files modified with the correct current experiment data directory. The default path in the config files will be a valid path allowing LabRecorder to still record data. (Example for Lion: "StorageLocation=/data/cat/LangLab/test_data/lion/eeg_fnirs_pupil/Lion_eeg_fnirs_pupil.xdf")