Adding "export_redcap_menu" & "say_announcement_menu" programs and their directories to the tomcat repository. (#486)
Rick Champlin: I am adding two programs I developed and their directories to git hub:
 "\tomcat\human_experiments\lab_software\export_redcap_data\export_redcap_menu"
and
"\tomcat\human_experiments\lab_software\say_announcements\say_announcement_menu"

Both of these programs have been used successfully multiple times for experiments. I feel it is now time to get them into the GitHub tomcat repository. 

export_redcap_menu:
Is a bash script program with menu to export REDCap Data to an Experiment Directory on "data/cat/".
The program is started in by "./export_redcap_menu".
It was designed to be a terminal only program with no "gui" interface so that it can be run remotely through ssh.
It can be run with a command line -h option to open it's help file with out opening the menu "./export_redcap_menu -h"

say_announcement_menu:
Is a bash script program with menu to play announcements into the participants earbuds during an experiment. The announcement, if enabled will also play on server Tom in the control room so that experiment operators can hear what the participants are hearing.
The program is started in by "./say_announcement_menu".
It too was designed to be a terminal only program with no "gui" interface so that it can be run remotely through ssh
and can be run with a command line -h option to open it's help file with out opening the menu
"./say_announcement_menu -h".
say_announcement_menu stores it's pre-set announcement files in:
"data/tom/LangLab/experiments/say_announcements/"
The pre-set and temp announcement files are stored in the above directory because all the Devices (Lion, Tiger, Leopard and Tom), have a "NFS" mapping to it. I have a included a backup directory with the current announcement files in
"say_announcements\announcement_files_bk". 

---------

Co-authored-by: Eric Duong <eduong@email.arizona.edu>